+++ tests/integration/test_mindful.py
@@ -30,6 +30,14 @@ class MockClass:
         return test
 
 
+class MockClass2:
+    status_code = 204
+    content = b""
+
+    def json():
+        return None
+
+
 @pytest.mark.init
 def test_instance_mindful():
     mf = Mindful(header=header)
@@ -67,6 +75,17 @@ def test_mindful_api_response3(mock_api_response):
     assert mf.endpoint == "responses"
 
 
+@mock.patch("viadot.sources.mindful.handle_api_response", return_value=MockClass)
+@pytest.mark.connect
+def test_mindful_api_response4(mock_api_response):
+    mf = Mindful(header=header)
+
+    response = mf.get_survey_list()
+
+    assert response.status_code == 200 and isinstance(response.json(), list)
+    assert mf.endpoint == "surveys"
+
+
 @mock.patch("viadot.sources.Mindful._mindful_api_response", return_value=MockClass)
 @pytest.mark.save
 def test_mindful_interactions(mock_connection):
@@ -89,3 +108,23 @@ def test_mindful_responses(mock_connection):
     assert mf.endpoint == "responses" and isinstance(mf.endpoint, str)
     assert os.path.exists("responses.csv")
     os.remove("responses.csv")
+
+
+@mock.patch("viadot.sources.Mindful._mindful_api_response", return_value=MockClass)
+@pytest.mark.save
+def test_mindful_surveys(mock_connection):
+    mf = Mindful(header=header)
+    response = mf.get_survey_list()
+    mf.response_to_file(response)
+
+    assert mf.endpoint == "surveys" and isinstance(mf.endpoint, str)
+    assert os.path.exists("surveys.csv")
+    os.remove("surveys.csv")
+
+
+@mock.patch("viadot.sources.Mindful._mindful_api_response", return_value=MockClass2)
+@pytest.mark.exception
+def test_file_exception(mock_mindful):
+    mf = MindfulToCSV()
+    response = mf.run(credentials_mindful=credentials_mindful)
+    assert response == None