+++ viadot/tasks/genesys.py
@@ -30,6 +30,7 @@ class GenesysToCSV(Task):
         report_url: str = None,
         report_columns: List[str] = None,
         credentials_genesys: Dict[str, Any] = None,
+        timeout: int = 3600,
         *args: List[Any],
         **kwargs: Dict[str, Any],
     ):
@@ -50,6 +51,8 @@ class GenesysToCSV(Task):
             schedule_id (str, optional): The ID of report. Defaults to None.
             report_url (str, optional): The url of report generated in json response. Defaults to None.
             report_columns (List[str], optional): List of exisiting column in report. Defaults to None.
+            timeout(int, optional): The amount of time (in seconds) to wait while running this task before
+                a timeout occurs. Defaults to 3600.
         """
 
         self.logger = prefect.context.get("logger")
@@ -69,6 +72,7 @@ class GenesysToCSV(Task):
 
         super().__init__(
             name=self.report_name,
+            timeout=timeout,
             *args,
             **kwargs,
         )
@@ -223,6 +227,7 @@ class GenesysToDF(Task):
         report_url: str = None,
         report_columns: List[str] = None,
         credentials_genesys: Dict[str, Any] = None,
+        timeout: int = 3600,
         *args: List[Any],
         **kwargs: Dict[str, Any],
     ):
@@ -237,6 +242,7 @@ class GenesysToDF(Task):
 
         super().__init__(
             name="genesys_to_df",
+            timeout=timeout,
             *args,
             **kwargs,
         )